Add firestore mock page for loading/unloading test data.
